products-documentation-ingestion/requirements.txt
beautifulsoup4==4.13.3
einops==0.8.1
html2text==2024.2.26
langchain==0.3.19
langchain-community==0.3.17
langchain-milvus==0.1.7
lxml==5.3.1
openai==1.63.2
pymilvus==2.4.10
pypdf==5.3.0
python-dotenv==1.0.1
tqdm==4.67.1
kfp==2.11.0